+++ aioxmpp/service.py
@@ -18,6 +18,7 @@ Both of these are provided in this module.
 import abc
 import asyncio
 import logging
+import warnings
 
 
 class Meta(abc.ABCMeta):
@@ -29,23 +30,52 @@ class Meta(abc.ABCMeta):
     Services have dependencies. A :class:`Meta` instance (i.e. a service class)
     can declare dependencies using the following two attributes.
 
-    .. attribute:: SERVICE_BEFORE
+    .. attribute:: ORDER_BEFORE
 
        An iterable of :class:`Service` classes before which the class which is
        currently being declared needs to be instanciated.
 
-       Thus, any service which occurs in :attr:`SERVICE_BEFORE` will be
+       Thus, any service which occurs in :attr:`ORDER_BEFORE` will be
        instanciated *after* this class (if at all).
+
+       .. versionadded:: 0.3
+
+    .. attribute:: SERVICE_BEFORE
+
+       Before 0.3, this was the name of the :attr:`ORDER_BEFORE` attribute. It
+       is still supported, but use emits a :data:`DeprecationWarning`. It must
+       not be mixed with :attr:`ORDER_BEFORE` or :attr:`ORDER_AFTER` on a class
+       declaration, or the declaration will raise :class:`ValueError`.
+
+       Support for this attribute will be removed in 1.x; starting with 1.0,
+       using this attribute will raise a :class:`TypeError` on class
+       declaration and a :class:`AttributeError` when accessing it on a class
+       or instance.
+
+       .. deprecated:: 0.3
+
+    .. attribute:: ORDER_AFTER
+
+       An iterable of :class:`Service` classes which would be instanciated
+       after the class which is currently being declared, if at all.
+
+       Classes which are declared in this attribute are not forced to be
+       instanciated (unlike with :attr:`ORDER_BEFORE`). However, if any of
+       these classes is requested, it is made sure that *this* class is
+       instanciated before.
+
+       .. newinversion:: 0.3
 
     .. attribute:: SERVICE_AFTER
 
-       An iterable of :class:`Service` classes which would be instanciated
-       after the class which is currently being declared, it at all.
+       Before 0.3, this was the name of the :attr:`ORDER_AFTER` attribute. It
+       is still supported, but use emits a :data:`DeprecationWarning`. It must
+       not be mixed with :attr:`ORDER_BEFORE` or :attr:`ORDER_AFTER` on a class
+       declaration, or the declaration will raise :class:`ValueError`.
 
-       Classes which are declared in this attribute are not forced to be
-       instanciated (unlike with :attr:`SERVICE_BEFORE`). However, if any of
-       these classes is requested, it is made sure that *this* class is
-       instanciated before.
+       See :attr:`SERVICE_BEFORE` for details on the deprecation cycle.
+
+       .. deprecated:: 0.3
 
     The dependencies are inherited from bases unless the `inherit_dependencies`
     keyword argument is set to false.
@@ -63,19 +93,33 @@ class Meta(abc.ABCMeta):
             pass
 
         class Bar(metaclass=service.Meta):
-            SERVICE_BEFORE = [Foo]
+            ORDER_BEFORE = [Foo]
 
         class Baz(metaclass=service.Meta):
-            SERVICE_BEFORE = [Bar]
+            ORDER_BEFORE = [Bar]
 
         class Fourth(metaclass=service.Meta):
-            SERVICE_BEFORE = [Bar]
+            ORDER_BEFORE = [Bar]
 
     ``Baz`` and ``Fourth`` will be instanciated before ``Bar`` and ``Bar`` will
     be instanciated before ``Foo``. There is no dependency relationship between
     ``Baz`` and ``Fourth``.
 
+    Inheritance works too::
 
+        class Foo(metaclass=service.Meta):
+            pass
+
+        class Bar(metaclass=service.Meta):
+            ORDER_BEFORE = [Foo]
+
+        class Baz(Bar):
+            # has ORDER_BEFORE == {Foo}
+            pass
+
+        class Fourth(Bar, inherit_dependencies=False):
+            # has empty ORDER_BEFORE
+            pass
 
     """
 
@@ -102,16 +146,32 @@ class Meta(abc.ABCMeta):
         return classes
 
     def __new__(mcls, name, bases, namespace, inherit_dependencies=True):
+        if "SERVICE_BEFORE" in namespace or "SERVICE_AFTER" in namespace:
+            if "ORDER_BEFORE" in namespace or "ORDER_AFTER" in namespace:
+                raise ValueError("declaration mixes old and new ordering "
+                                 "attribute names (SERVICE_* vs. ORDER_*)")
+            warnings.warn(
+                "SERVICE_BEFORE/AFTER used on class; use ORDER_BEFORE/AFTER",
+                DeprecationWarning)
+            try:
+                namespace["ORDER_BEFORE"] = namespace.pop("SERVICE_BEFORE")
+            except KeyError:
+                pass
+            try:
+                namespace["ORDER_AFTER"] = namespace.pop("SERVICE_AFTER")
+            except KeyError:
+                pass
+
         before_classes = mcls.collect_and_inherit(
             bases,
             namespace,
-            "SERVICE_BEFORE",
+            "ORDER_BEFORE",
             inherit_dependencies)
 
         after_classes = mcls.collect_and_inherit(
             bases,
             namespace,
-            "SERVICE_AFTER",
+            "ORDER_AFTER",
             inherit_dependencies)
 
         if before_classes & after_classes:
@@ -120,20 +180,22 @@ class Meta(abc.ABCMeta):
                 next(iter(before_classes & after_classes)).__qualname__
             ))
 
-        namespace["SERVICE_BEFORE"] = set(before_classes)
-        namespace["SERVICE_AFTER"] = set(after_classes)
+        namespace["ORDER_BEFORE"] = set(before_classes)
+        namespace["ORDER_AFTER"] = set(after_classes)
 
         return super().__new__(mcls, name, bases, namespace)
 
     def __init__(self, name, bases, namespace, inherit_dependencies=True):
         super().__init__(name, bases, namespace)
-        for cls in self.SERVICE_BEFORE:
-            cls.SERVICE_AFTER.add(self)
-        for cls in self.SERVICE_AFTER:
-            cls.SERVICE_BEFORE.add(self)
+        for cls in self.ORDER_BEFORE:
+            cls.ORDER_AFTER.add(self)
+        for cls in self.ORDER_AFTER:
+            cls.ORDER_BEFORE.add(self)
+        self.SERVICE_BEFORE = self.ORDER_BEFORE
+        self.SERVICE_AFTER = self.ORDER_AFTER
 
     def __lt__(self, other):
-        return other in self.SERVICE_BEFORE
+        return other in self.ORDER_BEFORE
 
     def __le__(self, other):
         return self < other
@@ -195,7 +257,8 @@ class Service(metaclass=Meta):
         Actual implementation of the shut down process.
 
         This *must* be called using super from inheriting classes after their
-        own shutdown procedure.
+        own shutdown procedure. Inheriting classes *must* override this method
+        instead of :meth:`shutdown`.
         """
 
     @asyncio.coroutine
@@ -211,9 +274,7 @@ class Service(metaclass=Meta):
 
            Developers creating subclasses of :class:`Service` to implement
            services should not override this method. Instead, they should
-           override the (intentionally undocumented) :meth:`_shutdown` method,
-           which is also a coroutine and has the same signature as
-           :meth:`shutdown`.
+           override the :meth:`_shutdown` method.
 
         """
         yield from self._shutdown()